requirements.txt
python-telegram-bot[job-queue]==21.9
pytz==2024.2
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
google-generativeai==0.8.3
